+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add `Say.clear_esc` method for clearing `^C` (`ctrl+c` output) from $stdout.
+
 ## [0.7.0] - 2025-1-4
 
 - Update minimum Ruby version from 3.1 -> 3.2